unicode_winapi: Fix params
Also add linting to the CI.

+++ pe-parser-library/src/unicode_winapi.cpp
@@ -22,20 +22,34 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#include <parser-library/to_string.h>
 #include <Windows.h>
+#include <parser-library/to_string.h>
 
 namespace peparse {
 std::string from_utf16(const UCharString &u) {
   std::string result;
-  std::size_t size = WideCharToMultiByte(CP_UTF8, 0, u.data(), u.size(), nullptr, 0, nullptr, nullptr);
+  std::size_t size = WideCharToMultiByte(CP_UTF8,
+                                         0,
+                                         u.data(),
+                                         static_cast<int>(u.size()),
+                                         nullptr,
+                                         0,
+                                         nullptr,
+                                         nullptr);
 
   if (size <= 0) {
     return result;
   }
 
   result.reserve(size);
-  WideCharToMultiByte(CP_UTF8, 0, u.data(), u.size(), &result[0], result.capacity(), nullptr, nullptr);
+  WideCharToMultiByte(CP_UTF8,
+                      0,
+                      u.data(),
+                      static_cast<int>(u.size()),
+                      &result[0],
+                      static_cast<int>(result.capacity()),
+                      nullptr,
+                      nullptr);
 
   return result;
 }
